refactor(odws): move chain pre-processing to chain/preprocess.ts
Extract preProcessForChain, prefixMessage, compactSizeEncode, uleb128Encode
from provider.ts (1695→1552 lines) into chain/preprocess.ts.

Provider now imports from chain/preprocess.js. Exported for direct use.
